pyproject.toml
[tool.poetry]
name = "wewu-less"
version = "0.1.0"
description = "WeWu cloud functions"
authors = ["Przemysław Fuchs <fuchs.przemyslaw@gmail.com>"]
license = "DWTFYW"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
marshmallow = "^3.20.1"
marshmallow-dataclass = "^8.6.0"
google-cloud-logging = "^3.9.0"
pymongo = "^4.6.1"
google-cloud-pubsub = "^2.19.0"
functions-framework = "^3.5.0"
structlog = "^23.3.0"
structlog-gcp = "^0.2.3"
testcontainers = "^3.7.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
